Move webservice File to package Data, update CategoryWSAdapter to work with GSON library to deserialize Object Need to test
