+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/configuration/MyQCMConstants.java
@@ -10,4 +10,5 @@ public class MyQCMConstants {
     public final static String CONST_URL_LOGIN = "http://192.168.1.16/MY_QCM/web/app_dev.php/api/userauths";
     public final static String CONST_VALUE_LOGIN = "login";
     public final static String CONST_VALUE_PWD = "password";
+    public final static String CONST_VALUE_ID_USER = "user_id";
 }
